-переносит типы ответов сервера в отдельный файл -переподключает типы
