+++ src/mocks/details.ts
@@ -1,19 +1,4 @@
-import { TOffer } from './offers';
-
-type THost = {
-  isPro: boolean;
-  name: string;
-  avatarUrl: string;
-};
-
-export type TDetail = Omit<TOffer, 'previewImage'> & {
-  description: string;
-  images: string[];
-  goods: string[];
-  host: THost;
-  bedrooms: number;
-  maxAdults: number;
-};
+import { TDetail } from '../types/details';
 
 const DETAILS: TDetail[] = [
   {
