+++ src/mocks/offers.ts
@@ -1,26 +1,4 @@
-type TCityLocation = {
-  latitude: number;
-  longitude: number;
-  zoom: number;
-};
-
-type TCity = {
-  name: string;
-  location: TCityLocation;
-};
-
-export type TOffer = {
-  id: string;
-  title: string;
-  type: string;
-  price: number;
-  previewImage: string;
-  city: TCity;
-  location: TCityLocation;
-  isFavorite: boolean;
-  isPremium: boolean;
-  rating: number;
-};
+import { TOffer } from '../types/offer';
 
 const OFFERS: TOffer[] = [
   {
