+++ src/mocks/reviews.ts
@@ -1,21 +1,4 @@
-type TUser = {
-  name: string;
-  avatarUrl: string;
-  isPro: boolean;
-};
-
-type TComments = {
-  id: string;
-  comment: string;
-  date: string;
-  rating: number;
-  user: TUser;
-};
-
-export type TReview = {
-  id: string;
-  comments: TComments[];
-};
+import { TReview } from '../types/review';
 
 const REVIEWS: TReview[] = [
   {
